+++ react-components/src/components/Table/index.tsx
@@ -1,16 +1,17 @@
-import React, { useState } from 'react';
+import React, { useState, useRef, useMemo } from 'react';
 import classNames from 'classnames';
 import {
   CellContext,
   createColumnHelper,
   flexRender,
   getCoreRowModel,
+  Row,
   useReactTable,
 } from '@tanstack/react-table';
+import { useVirtualizer } from '@tanstack/react-virtual';
 
 import { TableProps } from './model';
 import TableRow from './components/RowTable';
-import { Spinner } from '../Spinner';
 import { FilterType } from './model/enum/filter-type.enum';
 
 import s from './style.module.scss';
@@ -19,44 +20,55 @@ export function Table<T>({
   dataSource,
   columns,
   sticky = false,
-  loading = false,
+  skeleton = null,
   dictionary = {},
+  cellParams = {
+    verticalClamp: 1,
+  },
+  height = '100vh',
+  onClick,
+  acrossLine = false,
   className,
   ...props
-}: TableProps<T>): JSX.Element {
+}: TableProps<T>): React.ReactElement {
   const columnHelper = createColumnHelper<T>();
-  const tableColumns = columns.map(
-    ({ name, caption, cellComponent, propertyName, filterType }) => {
-      return columnHelper.accessor(name as any, {
-        id: name,
-        cell: (info: CellContext<T, number>) => {
-          const infoValue = info.getValue();
-          let dictValue = '';
+  const tableColumns = useMemo(() => {
+    return columns.map(
+      ({ name, caption, cellComponent, propertyName, filterType, width }) => {
+        return columnHelper.accessor(name as any, {
+          id: name,
+          cell: (info: CellContext<T, number>) => {
+            const infoValue = info.getValue();
+            let dictValue = '';
 
-          if (filterType === FilterType.list && propertyName) {
-            dictValue =
-              dictionary[propertyName]?.find(({ id }) => id === infoValue)
-                ?.name ?? '';
-          }
+            if (filterType === FilterType.list && propertyName) {
+              dictValue =
+                dictionary[propertyName]?.find(({ id }) => id === infoValue)
+                  ?.name ?? '';
+            }
 
-          const value = dictValue || infoValue;
+            const value = dictValue || infoValue;
 
-          if (cellComponent) {
-            return React.createElement(cellComponent, {
-              value,
-            });
-          }
+            let result;
 
-          if (typeof value === 'object' && value !== null) {
-            return Object.values(value).join(' — ');
-          }
+            if (cellComponent) {
+              result = React.createElement(cellComponent, {
+                value,
+              });
+            } else if (typeof value === 'object' && value !== null) {
+              result = JSON.stringify(value);
+            } else {
+              result = value;
+            }
 
-          return value;
-        },
-        header: () => caption,
-      });
-    },
-  );
+            return <span>{result}</span>;
+          },
+          header: () => caption,
+          size: width,
+        });
+      },
+    );
+  }, [columns, dictionary, columnHelper]);
 
   const [rowSelection, setRowSelection] = useState({});
 
@@ -72,59 +84,69 @@ export function Table<T>({
     getCoreRowModel: getCoreRowModel(),
   });
 
-  if (loading) {
-    const loadingRows: string[] = [];
-    for (let i = 1; i <= 20; i++) {
-      loadingRows.push(String(i));
-    }
+  const parentRef = useRef<HTMLDivElement>(null);
+  const { rows } = table.getRowModel();
+  const virtualizer = useVirtualizer({
+    count: rows.length,
+    paddingStart: parentRef.current?.querySelector('thead')?.offsetHeight ?? 28,
+    getScrollElement: () => parentRef.current,
+    estimateSize: () => 28,
+    overscan: 5,
+  });
 
-    return (
-      <div className={s.loadTable}>
-        <table {...props} className={classNames(s.table, className)}>
-          <tbody>
-            {loadingRows.map((item) => (
-              <tr key={item}>
-                <td></td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
+  const cellStyles = {
+    '--cell-vert-clamp': cellParams.verticalClamp,
+    '--tbody-transform': `${virtualizer.getVirtualItems()[0]?.start}px`,
+  };
 
-        <div className={s.loadTableSpinner}>
-          <Spinner color="var(--color-primary-50)" size={20} />
-        </div>
-      </div>
-    );
-  }
+  if (skeleton) return skeleton;
 
   return (
-    <table {...props} className={classNames(s.table, className)}>
-      <thead className={classNames(sticky && s.sticky)}>
-        {table.getHeaderGroups().map((headerGroup) => (
-          <tr key={headerGroup.id}>
-            {headerGroup.headers.map((header) => (
-              <th key={header.id}>
-                {header.isPlaceholder
-                  ? null
-                  : flexRender(
-                      header.column.columnDef.header,
-                      header.getContext(),
-                    )}
-              </th>
-            ))}
-          </tr>
-        ))}
-      </thead>
+    <div className={s.root} ref={parentRef} style={{ height }}>
+      <table
+        {...props}
+        className={classNames(s.table, className)}
+        style={{
+          ...cellStyles,
+          height: virtualizer.getTotalSize(),
+        }}
+      >
+        <thead className={classNames(sticky && s.sticky)}>
+          {table.getHeaderGroups().map((headerGroup) => (
+            <tr key={headerGroup.id}>
+              {headerGroup.headers.map((header) => (
+                <th key={header.id} style={{ width: header.getSize() }}>
+                  {header.isPlaceholder
+                    ? null
+                    : flexRender(
+                        header.column.columnDef.header,
+                        header.getContext(),
+                      )}
+                </th>
+              ))}
+            </tr>
+          ))}
+        </thead>
 
-      <tbody>
-        {table.getRowModel().rows.map((row) => (
-          <TableRow
-            key={row.id}
-            row={row}
-            isSelectedRow={row.getIsSelected()}
-          />
-        ))}
-      </tbody>
-    </table>
+        <tbody>
+          {virtualizer.getVirtualItems().map((virtualRow) => {
+            const row = rows[virtualRow.index] as Row<T>;
+
+            return (
+              <TableRow
+                key={virtualRow.key}
+                virtualIndex={virtualRow.index}
+                rowRef={virtualizer.measureElement}
+                row={row}
+                columns={columns}
+                isSelectedRow={row.getIsSelected()}
+                onClick={onClick}
+                acrossLine={acrossLine}
+              />
+            );
+          })}
+        </tbody>
+      </table>
+    </div>
   );
 }