+++ react-components/src/components/Table/model/index.ts
@@ -1,10 +1,14 @@
-import { HTMLAttributes } from 'react';
+import React, { HTMLAttributes } from 'react';
+
 import { TableColumn } from './table-column';
-import { IDictionary } from 'tetacom/react-components';
+import { IDictionary } from './dictionary';
+import { CellParamsType } from './cell-params';
+import { ICellInstance } from './i-cell-instance';
 
 export type TableRef = HTMLTableElement;
 
-export interface TableProps<T> extends HTMLAttributes<TableRef> {
+export interface TableProps<T>
+  extends Omit<HTMLAttributes<TableRef>, 'onClick'> {
   // Массив записей данных для отображения
   dataSource: Array<T>;
 
@@ -14,9 +18,21 @@ export interface TableProps<T> extends HTMLAttributes<TableRef> {
   // Зафиксировать шапку при скролле
   sticky?: boolean;
 
-  // Установить статус загрузки таблицы
-  loading?: boolean;
+  // Установить заполнитель таблицы
+  skeleton?: React.ReactElement | null;
 
   // Словарь
   dictionary?: IDictionary;
+
+  // Настройка ячейки
+  cellParams?: CellParamsType;
+
+  // Высота таблицы
+  height?: React.CSSProperties['height'];
+
+  // Событие клика по строке
+  onClick?: (cell: ICellInstance<T>) => void;
+
+  // Выделять строчки таблицы через одну
+  acrossLine?: boolean;
 }