+++ react-components/src/components/Tabs/components/Highlight/index.tsx
@@ -7,6 +7,8 @@ import { Direction } from '../../model';
 
 import s from './style.module.scss';
 
+const SIZE = 2;
+
 interface Props {
   tabsRef: HTMLUListElement | null;
   selectedKey: TabType['key'];
@@ -40,14 +42,14 @@ export const Highlight: FC<Props> = ({
 
   const horizontal = {
     x: offsetWidth,
-    y: 0,
+    y: -1 * SIZE,
     width: currentTabWidth,
-    height: '100%',
+    height: SIZE,
   };
   const vertical = {
     x: 0,
     y: offsetHeight,
-    width: '100%',
+    width: SIZE,
     height: currentTabHeight,
   };