+++ react-components/src/components/Typography/model/public-api.ts
@@ -1,0 +1,1 @@
+export type { FontVariant as FontVariantType } from './index';